# currency: boost retrieval speed when calling currency.NewCode()