@thrasher- 2022-08-31
tACK, quite a significant speed boost. Nice stuff!

```
PR:
goos: windows
goarch: amd64
pkg: github.com/thrasher-corp/gocryptotrader/currency
cpu: AMD Ryzen Threadripper 3970X 32-Core Processor
BenchmarkNewCode
BenchmarkNewCode-64
14170744	        77.24 ns/op	       8 B/op	       1 allocs/op

main:
goos: windows
goarch: amd64
pkg: github.com/thrasher-corp/gocryptotrader/currency
cpu: AMD Ryzen Threadripper 3970X 32-Core Processor
BenchmarkNewCode
BenchmarkNewCode-64 313797	      3364 ns/op	       8 B/op	       1 allocs/op

PR:
goos: linux
goarch: amd64
pkg: github.com/thrasher-corp/gocryptotrader/currency
cpu: AMD Ryzen Threadripper 3970X 32-Core Processor
BenchmarkNewCode
BenchmarkNewCode-64    	13845956	        84.42 ns/op	       8 B/op	       1 allocs/op

main:
goos: linux
goarch: amd64
pkg: github.com/thrasher-corp/gocryptotrader/currency
cpu: AMD Ryzen Threadripper 3970X 32-Core Processor
BenchmarkNewCode
BenchmarkNewCode-64    	  310576	      3759 ns/op	       8 B/op	       1 allocs/op

```